+++ R/utilities.R
@@ -1,0 +1,29 @@
+#'  practice1 function 
+#' 
+#'  @param dir directory (under projet root) to compress
+#'  @return 
+#'  @export
+
+#'  
+#'
+practice1  <- function(dir) {
+	1}
+
+practice1()
+
+
+
+
+
+#'  practice function 
+#' 
+#'  
+#'  @return 1
+#'  @export
+
+#'  
+#'
+practice  <- function() {
+	1}
+
+practice()